version 16.4

+++ test/update_test_status.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+export PATH=/usr/local/bin:/usr/bin:/bin:/usr/local/sbin:/usr/sbin:/sbin
+
 set -e
 set -x
 
@@ -11,7 +13,9 @@ exec 2>&1
 cd ~/tint2
 git reset --hard
 git pull
-
+last=$(cat .last-reg-test || true)
+curr=$(git rev-parse --verify HEAD)
+[ "$last" == "$curr" ] && exit 0
 
 cd ~/tint2.wiki
 git reset --hard
@@ -27,3 +31,6 @@ cd ~/tint2.wiki
 git add tests.md
 git commit -am 'Update test results'
 git push origin master
+
+cd ~/tint2
+echo "$curr" > .last-reg-test
